+++ base/src/main/java/proguard/analysis/cpa/domain/taint/TaintSink.java
@@ -18,18 +18,21 @@
 
 package proguard.analysis.cpa.domain.taint;
 
+import java.util.Objects;
+import java.util.function.Predicate;
 import proguard.classfile.Signature;
 
 /**
  * A {@link TaintSink} specifies a sink for the taint analysis. A sink can be sensitive to
  * the instance, the arguments, or the static fields. If a sink S is sensitive to X, then
  * if X is tainted, we conclude that the taint has reached S.
- *
- * @author Dmitry Ivanov
  */
 public abstract class TaintSink
 {
-    public final Signature signature;
+
+    protected static final Predicate<TaintSource> IS_VALID_FOR_SOURCE_DEFAULT = x -> true;
+    public final           Signature              signature;
+    public final           Predicate<TaintSource> isValidForSource;
 
     /**
      * Create a taint sink.
@@ -38,14 +41,41 @@ public abstract class TaintSink
      */
     public TaintSink(Signature signature)
     {
+        this(signature, IS_VALID_FOR_SOURCE_DEFAULT);
+    }
+
+    /**
+     * Create a taint sink.
+     *
+     * @param signature        the signature of a sink method
+     * @param isValidForSource predicate on whether the sink is valid for a given source
+     */
+    public TaintSink(Signature signature, Predicate<TaintSource> isValidForSource)
+    {
         this.signature = signature;
+        this.isValidForSource = isValidForSource;
     }
 
     @Override
-    public abstract boolean equals(Object o);
+    public boolean equals(Object other)
+    {
+        if (this == other)
+        {
+            return true;
+        }
+        if (!(other instanceof TaintSink))
+        {
+            return false;
+        }
+        TaintSink taintSink = (TaintSink) other;
+        return Objects.equals(signature, taintSink.signature) && Objects.equals(isValidForSource, taintSink.isValidForSource);
+    }
 
     @Override
-    public abstract int hashCode();
+    public int hashCode()
+    {
+        return Objects.hash(signature, isValidForSource);
+    }
 
     @Override
     public abstract String toString();