+++ base/src/main/java/proguard/analysis/cpa/jvm/domain/taint/JvmInvokeTaintSink.java
@@ -20,10 +20,10 @@ package proguard.analysis.cpa.jvm.domain.taint;
 
 import java.util.HashSet;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.Set;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
+import proguard.analysis.cpa.domain.taint.TaintSource;
 import proguard.analysis.cpa.interfaces.CallEdge;
 import proguard.analysis.cpa.jvm.cfa.edges.JvmCallCfaEdge;
 import proguard.analysis.cpa.jvm.cfa.edges.JvmCfaEdge;
@@ -38,18 +38,16 @@ import proguard.classfile.util.ClassUtil;
  * A {@link JvmTaintSink} on a method invocation.
  * This sinks can be sensitive to the instance, the arguments, or the static fields. If a sink S is sensitive to X, then
  * if X is tainted, we conclude that the taint has reached S.
- *
- * @author Dmitry Ivanov
  */
 public class JvmInvokeTaintSink
     extends JvmTaintSink
 {
 
-    public final Optional<Predicate<Call>> callMatcher;
-
-    public final boolean      takesInstance;
-    public final Set<Integer> takesArgs;
-    public final Set<String>  takesGlobals;
+    private static final Predicate<Call> CALL_MATCHER_DEFAULT = x -> true;
+    public final         boolean         takesInstance;
+    public final         Set<Integer>    takesArgs;
+    public final         Set<String>     takesGlobals;
+    public final         Predicate<Call> callMatcher;
 
     /**
      * Create a taint sink.
@@ -59,9 +57,15 @@ public class JvmInvokeTaintSink
      * @param takesArgs     a set of sensitive arguments
      * @param takesGlobals  a set of sensitive global variables
      */
+    @Deprecated
     public JvmInvokeTaintSink(Signature signature, boolean takesInstance, Set<Integer> takesArgs, Set<String> takesGlobals)
     {
-        this(signature, Optional.empty(), takesInstance, takesArgs, takesGlobals);
+        this(signature,
+             IS_VALID_FOR_SOURCE_DEFAULT,
+             takesInstance,
+             takesArgs,
+             takesGlobals,
+             CALL_MATCHER_DEFAULT);
     }
 
     /**
@@ -73,33 +77,131 @@ public class JvmInvokeTaintSink
      * @param takesArgs     a set of sensitive arguments
      * @param takesGlobals  a set of sensitive global variables
      */
-    public JvmInvokeTaintSink(Signature signature, Predicate<Call> callMatcher, boolean takesInstance, Set<Integer> takesArgs, Set<String> takesGlobals)
+    @Deprecated
+    public JvmInvokeTaintSink(Signature       signature,
+                              Predicate<Call> callMatcher,
+                              boolean         takesInstance,
+                              Set<Integer>    takesArgs,
+                              Set<String>     takesGlobals)
     {
-        this(signature, Optional.of(callMatcher), takesInstance, takesArgs, takesGlobals);
+        this(signature,
+             IS_VALID_FOR_SOURCE_DEFAULT,
+             takesInstance,
+             takesArgs,
+             takesGlobals,
+             callMatcher);
     }
 
     /**
      * Create a taint sink.
      *
-     * @param signature     the signature of a sink method
-     * @param callMatcher   an optional predicate on whether the call matches this taint sink
-     * @param takesInstance whether the sink is sensitive to the calling instance
-     * @param takesArgs     a set of sensitive arguments
-     * @param takesGlobals  a set of sensitive global variables
+     * @param signature        the signature of a sink method
+     * @param isValidForSource predicate on whether the sink is valid for a given source
+     * @param takesInstance    whether the sink is sensitive to the calling instance
+     * @param takesArgs        a set of sensitive arguments
+     * @param takesGlobals     a set of sensitive global variables
+     * @param callMatcher      predicate on whether a call matches this taint sink
      */
-    private JvmInvokeTaintSink(Signature signature, Optional<Predicate<Call>> callMatcher, boolean takesInstance, Set<Integer> takesArgs, Set<String> takesGlobals)
+    protected JvmInvokeTaintSink(Signature              signature,
+                                 Predicate<TaintSource> isValidForSource,
+                                 boolean                takesInstance,
+                                 Set<Integer>           takesArgs,
+                                 Set<String>            takesGlobals,
+                                 Predicate<Call>        callMatcher)
     {
-        super(signature);
+        super(signature, isValidForSource);
 
         if (!takesInstance && takesArgs.isEmpty() && takesGlobals.isEmpty())
         {
             throw new RuntimeException(String.format("Tainted sink for method %s must have taint somewhere!", signature));
         }
 
-        this.callMatcher = callMatcher;
         this.takesInstance = takesInstance;
         this.takesArgs = takesArgs;
         this.takesGlobals = takesGlobals;
+        this.callMatcher = callMatcher;
+    }
+
+    /**
+     * Builder for {@link proguard.analysis.cpa.jvm.domain.taint.JvmInvokeTaintSink}.
+     */
+    public static class Builder
+    {
+
+        protected       Predicate<TaintSource> isValidForSource    = IS_VALID_FOR_SOURCE_DEFAULT;
+        protected       Predicate<Call>        callMatcher         = CALL_MATCHER_DEFAULT;
+        protected       boolean                takesInstance       = false;
+        protected       Set<Integer>           takesArgs           = new HashSet<>();
+        protected       Set<String>            takesGlobals        = new HashSet<>();
+        protected final Signature              signature;
+
+        /**
+         * Create a new builder for {@link proguard.analysis.cpa.jvm.domain.taint.JvmInvokeTaintSink}.
+         *
+         * @param signature @param signature the signature of a sink method
+         */
+        public Builder(Signature signature)
+        {
+            this.signature = signature;
+        }
+
+        /**
+         * Build a {@link proguard.analysis.cpa.jvm.domain.taint.JvmInvokeTaintSink}.
+         */
+        public JvmInvokeTaintSink build()
+        {
+            return new JvmInvokeTaintSink(signature,
+                                          isValidForSource,
+                                          takesInstance,
+                                          takesArgs,
+                                          takesGlobals,
+                                          callMatcher);
+        }
+
+        /**
+         * @param isValidForSource predicate on whether the sink is valid for a given source
+         */
+        public Builder setIsValidForSource(Predicate<TaintSource> isValidForSource)
+        {
+            this.isValidForSource = isValidForSource;
+            return this;
+        }
+
+        /**
+         * @param takesInstance whether the sink is sensitive to the calling instance
+         */
+        public Builder setTakesInstance(boolean takesInstance)
+        {
+            this.takesInstance = takesInstance;
+            return this;
+        }
+
+        /**
+         * @param takesArgs a set of sensitive arguments
+         */
+        public Builder setTakesArgs(Set<Integer> takesArgs)
+        {
+            this.takesArgs = takesArgs;
+            return this;
+        }
+
+        /**
+         * @param takesGlobals a set of sensitive global variables
+         */
+        public Builder setTakesGlobals(Set<String> takesGlobals)
+        {
+            this.takesGlobals = takesGlobals;
+            return this;
+        }
+
+        /**
+         * @param callMatcher predicate on whether a call matches this taint sink
+         */
+        public Builder setCallMatcher(Predicate<Call> callMatcher)
+        {
+            this.callMatcher = callMatcher;
+            return this;
+        }
     }
 
     /**
@@ -132,7 +234,7 @@ public class JvmInvokeTaintSink
             return false;
         }
         CallEdge callEdge = (CallEdge) edge;
-        return signature.equals(callEdge.getCall().getTarget()) && callMatcher.map(p -> p.test(callEdge.getCall())).orElse(true);
+        return signature.equals(callEdge.getCall().getTarget()) && callMatcher.test(callEdge.getCall());
     }
 
     @Override
@@ -146,9 +248,12 @@ public class JvmInvokeTaintSink
         {
             return false;
         }
+        if (!super.equals(o))
+        {
+            return false;
+        }
         JvmInvokeTaintSink taintSink = (JvmInvokeTaintSink) o;
         return takesInstance == taintSink.takesInstance
-               && Objects.equals(signature, taintSink.signature)
                && Objects.equals(takesArgs, taintSink.takesArgs)
                && Objects.equals(takesGlobals, taintSink.takesGlobals)
                && Objects.equals(callMatcher, taintSink.callMatcher);
@@ -157,7 +262,7 @@ public class JvmInvokeTaintSink
     @Override
     public int hashCode()
     {
-        return Objects.hash(signature, takesInstance, takesArgs, takesGlobals, callMatcher);
+        return Objects.hash(super.hashCode(), takesInstance, takesArgs, takesGlobals, callMatcher);
     }
 
     @Override
@@ -185,7 +290,15 @@ public class JvmInvokeTaintSink
                                       .collect(Collectors.joining(", ")))
                   .append(")");
         }
-        callMatcher.ifPresent(p -> result.append(" filtered by ").append(p));
+
+        if (!IS_VALID_FOR_SOURCE_DEFAULT.equals(isValidForSource))
+        {
+            result.append(", filtered by source ").append(isValidForSource);
+        }
+        if (!CALL_MATCHER_DEFAULT.equals(callMatcher))
+        {
+            result.append(", filtered by call ").append(callMatcher);
+        }
 
         return result.toString();
     }