+++ base/src/main/java/proguard/analysis/cpa/jvm/domain/taint/JvmTaintSink.java
@@ -22,6 +22,8 @@ import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
+import java.util.function.Predicate;
+import proguard.analysis.cpa.domain.taint.TaintSource;
 import proguard.analysis.cpa.jvm.cfa.edges.JvmCfaEdge;
 import proguard.analysis.cpa.domain.taint.TaintSink;
 import proguard.analysis.cpa.jvm.witness.JvmMemoryLocation;
@@ -30,8 +32,6 @@ import proguard.classfile.Signature;
 /**
  * The {@link JvmTaintSink} adds an interface for extracting sensitive JVM memory locations and to check
  * if the sink matches a given cfa edge.
- *
- * @author Dmitry Ivanov
  */
 public abstract class JvmTaintSink
     extends TaintSink
@@ -40,6 +40,11 @@ public abstract class JvmTaintSink
     public JvmTaintSink(Signature signature)
     {
         super(signature);
+    }
+
+    public JvmTaintSink(Signature signature, Predicate<TaintSource> isValidForSource)
+    {
+        super(signature, isValidForSource);
     }
 
     /**
@@ -52,6 +57,11 @@ public abstract class JvmTaintSink
      */
     public abstract boolean matchCfaEdge(JvmCfaEdge edge);
 
+    /**
+     * Helper method taking a collection of sinks and converting it to a mapping that associates each sink
+     * with the memory locations which triggers it. For convenience the sinks are further grouped
+     * by their method signature.
+     */
     public static Map<Signature, Map<JvmTaintSink, Set<JvmMemoryLocation>>> convertSinksToMemoryLocations(Collection<? extends JvmTaintSink> taintSinks)
     {
         Map<Signature, Map<JvmTaintSink, Set<JvmMemoryLocation>>> result = new HashMap<>();
@@ -61,5 +71,17 @@ public abstract class JvmTaintSink
             result.computeIfAbsent(taintSink.signature, x -> new HashMap<>()).put(taintSink, memoryLocations);
         }
         return result;
+    }
+
+    @Override
+    public boolean equals(Object other)
+    {
+        return super.equals(other);
+    }
+
+    @Override
+    public int hashCode()
+    {
+        return super.hashCode();
     }
 }